Made bide open files by default, and only create files if they don't exist.

+++ src/bide.sh
@@ -1,14 +1,13 @@
 #! /bin/bash
 
 # Global Variables
-
 ## Directories
 bidedir=$HOME/.bide
 ftdir=$bidedir/fttemplate
 deftemplate=$ftdir/default.txt
+editor=$VISUAL
 
 # Functions
-
 ## Set up the bide directory structure
 ## if it does not exist.
 install() {
@@ -25,6 +24,10 @@ install() {
 	if [ ! -e $deftemplate ]; then
 		touch $deftemplate
 	fi
+
+	if [ -z $VISUAL ]; then
+		editor=/usr/bin/vim
+	fi
 }
 
 ## Check for correct formating of the
@@ -37,36 +40,46 @@ checkUsage() {
 }
 
 ## Creates a file based upon the name
-## passed in.
+## passed in. If the file already exits
+## it does nothing.
 createFile() {
 	# Get the extension and create the
 	# default file name.
 	filename=$1
-	ext=${filename##*.}
-	template=$ftdir/default.${ext}
 
-	# Check the case where there is no extension.
-	if [ $ext = $filename ]; then
-		template=$deftemplate
+	if [ ! -e $filename ]; then
+		ext=${filename##*.}
+		template=$ftdir/default.${ext}
+
+		# Check the case where there is no extension.
+		if [ $ext = $filename ]; then
+			template=$deftemplate
+		fi
+
+		# Check if the extension template file exits.
+		if [ ! -e $template ]; then
+			template=$deftemplate
+		fi
+
+		# Copy the template file.
+		cp $template $filename
 	fi
+}
 
-	# Check if the extension template file exits.
-	if [ ! -e $template ]; then
-		template=$deftemplate
-	fi
-
-	# Copy the template file.
-	cp $template $filename
+## Opens the system editor with the file
+## created.
+openEditor() {
+	$editor $@
 }
 
 ## Entry point of the program.
 main() {
 	checkUsage $@
 	install
-
 	for file in $@; do
 		createFile $file
 	done
+	openEditor $@
 }
 
 # Call the main method.
